requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
streamlit>=1.25.0
opencv-python>=4.6.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.0.0
Pillow>=9.0.0
